requirements.txt
# DJANGO
Django==5.1.6
djangorestframework==3.15.2
django-cleanup==9.0.0
django-crum==0.7.9
django-environ==0.12.0
django-user-agents==0.4.0
django-widget-tweaks==1.5.0
django-simple-history==3.8.0
django-cors-headers==4.7.0
django-imagekit==5.0.0
django-import-export==4.3.5
django-password-validators==1.7.3
requests==2.32.3
django-colorfield==0.12.0

# PostgreSQL
psycopg2-binary==2.9.10
# MySQL
django-mysql==4.16.0
mysqlclient==2.2.7
PyMySQL==1.1.1

# Images
Pillow==11.1.0
piexif==1.1.3

# Maps
django-leaflet==0.31.0
googlemaps==4.10.0

# Testing
django-debug-toolbar==4.3.0
ipython==8.32.0

# Deployment
whitenoise==6.8.2
uvicorn==0.34.0
gunicorn==23.0.0
weasyprint==64.0

# Emails
resend==2.19.0